backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pymupdf==1.25.4
httpx==0.28.1
beautifulsoup4==4.12.3
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
langgraph==1.0.10
litellm>=1.40.0
langchain-core==1.2.18
pymongo[srv]==4.12.1
python-json-logger==3.3.0
rapidfuzz==3.14.3
